+++ selection/randomized/randomization.py
@@ -12,13 +12,13 @@ from scipy.stats import laplace, norm as ndist
 
 class randomization(rr.smooth_atom):
 
-    def __init__(self, shape, density, grad_negative_log_density, sampler):
+    def __init__(self, shape, density, grad_negative_log_density, sampler, lipschitz=1):
         rr.smooth_atom.__init__(self,
                                 shape)
         self._density = density
         self._grad_negative_log_density = grad_negative_log_density
         self._sampler = sampler
-        self.lipschitz = 1.
+        self.lipschitz = lipschitz
 
     def smooth_objective(self, perturbation, mode='both', check_feasibility=False):
         """
@@ -58,7 +58,7 @@ class randomization(rr.smooth_atom):
         density = lambda x: rv.pdf(x)
         grad_negative_log_density = lambda x: x / scale**2
         sampler = lambda size: rv.rvs(size=shape + size)
-        return randomization(shape, density, grad_negative_log_density, sampler)
+        return randomization(shape, density, grad_negative_log_density, sampler, lipschitz=1./scale**2)
 
     @staticmethod
     def gaussian(covariance):
@@ -70,7 +70,7 @@ class randomization(rr.smooth_atom):
         density = lambda x: np.exp(-(x * precision.dot(x)).sum() / 2) / _const
         grad_negative_log_density = lambda x: precision.dot(x)
         sampler = lambda size: sqrt_precision.dot(np.random.standard_normal((p,) + size))
-        return randomization((p,), density, grad_negative_log_density, sampler)
+        return randomization((p,), density, grad_negative_log_density, sampler, lipschitz=np.linalg.svd(precision)[1].max())
 
     @staticmethod
     def laplace(shape, scale):
@@ -78,7 +78,7 @@ class randomization(rr.smooth_atom):
         density = lambda x: rv.pdf(x)
         grad_negative_log_density = lambda x: np.sign(x) / scale
         sampler = lambda size: rv.rvs(size=shape + size)
-        return randomization(shape, density, grad_negative_log_density, sampler)
+        return randomization(shape, density, grad_negative_log_density, sampler, lipschitz=1./scale**2)
 
     @staticmethod
     def logistic(shape, scale):
@@ -88,4 +88,4 @@ class randomization(rr.smooth_atom):
         # x/s + log(s) + 2 \log (1 + e(-x/s))
         grad_negative_log_density = lambda x: (1 - np.exp(-x / scale)) / ((1 + np.exp(-x / scale)) * scale)
         sampler = lambda size: np.random.logistic(loc=0, scale=scale, size=shape + size)
-        return randomization(shape, density, grad_negative_log_density, sampler)
+        return randomization(shape, density, grad_negative_log_density, sampler, lipschitz=.25/scale**2)
