+++ selection/randomized/tests/test_Mest.py
@@ -487,125 +487,3 @@ def test_logistic_saturated_active_coordinate(seed=None):
         pval = target_sampler.hypothesis_test(test_stat, active_observed, burnin=10000, ndraw=10000) # twosided by default
         return pval
 
-@wait_for_return_value
-def test_logistic_many_targets():
-    s, n, p = 5, 200, 20 
-
-    randomizer = randomization.laplace((p,), scale=1.)
-    X, y, beta, _ = logistic_instance(n=n, p=p, s=s, rho=0.1, snr=14)
-
-    nonzero = np.where(beta)[0]
-    lam_frac = 1.
-
-    loss = rr.glm.logistic(X, y)
-    epsilon = 1.
-
-    lam = lam_frac * np.mean(np.fabs(np.dot(X.T, np.random.binomial(1, 1. / 2, (n, 10000)))).max(0))
-    W = np.ones(p)*lam
-    penalty = rr.group_lasso(np.arange(p),
-                             weights=dict(zip(np.arange(p), W)), lagrange=1.)
-
-    print lam
-    M_est = glm_group_lasso(loss, epsilon, penalty, randomizer)
-
-    mv = multiple_views([M_est])
-    mv.solve()
-
-    active = M_est.overall
-    nactive = active.sum()
-
-    global_scaling = np.linalg.svd(X)[1].max()**2
-
-    def crude_target_scaling(_target_sampler):
-        result = np.linalg.svd(_target_sampler.target_inv_cov)[1].max()
-        for transform, objective in zip(_target_sampler.target_transform, _target_sampler.objectives):
-            result += np.linalg.svd(transform[0])[1].max()**2 * objective.randomization.lipschitz
-            result += np.linalg.svd(objective.score_transform[0])[1].max()**2 * objective.randomization.lipschitz
-        return result
-
-    if set(nonzero).issubset(np.nonzero(active)[0]):
-
-        pvalues = []
-        active_set = np.nonzero(active)[0]
-        inactive_selected = I = [i for i in np.arange(active_set.shape[0]) if active_set[i] not in nonzero]
-        active_selected = A = [i for i in np.arange(active_set.shape[0]) if active_set[i] in nonzero]
-
-        if not I:
-            return None
-        idx = I[0]
-        boot_target, target_observed = pairs_bootstrap_glm(loss, active, inactive=M_est.inactive)
-
-        sampler = lambda : np.random.choice(n, size=(n,), replace=True)
-        mv.setup_sampler(sampler, scaling=global_scaling)
-
-        # null saturated
-
-        def null_target(indices):
-            result = boot_target(indices)
-            return result[idx]
-
-        null_observed = np.zeros(1)
-        null_observed[0] = target_observed[idx]
-
-        target_sampler = mv.setup_target(null_target, null_observed)
-        #target_scaling = 5 * np.linalg.svd(target_sampler.target_transform[0][0])[1].max()**2# should have something do with noise scale too
-
-        test_stat = lambda x: x[0]
-        pval = target_sampler.hypothesis_test(test_stat, null_observed, burnin=10000, ndraw=8000, stepsize=1./crude_target_scaling(target_sampler)) # twosided by default
-        pvalues.append(pval)
-
-        # null selected
-
-        def null_target(indices):
-            result = boot_target(indices)
-            return np.hstack([result[idx], result[nactive:]])
-
-        null_observed = np.zeros_like(null_target(range(n)))
-        null_observed[0] = target_observed[idx]
-        null_observed[1:] = target_observed[nactive:]
-
-        target_sampler = mv.setup_target(null_target, null_observed, target_set=[0])
-        target_scaling = 5 * np.linalg.svd(target_sampler.target_transform[0][0])[1].max()**2# should have something do with noise scale too
-
-        test_stat = lambda x: x[0]
-        pval = target_sampler.hypothesis_test(test_stat, null_observed, burnin=10000, ndraw=8000, stepsize=1./crude_target_scaling(target_sampler)) # twosided by default
-        pvalues.append(pval)
-
-        # true saturated
-
-        idx = A[0]
-
-        def active_target(indices):
-            result = boot_target(indices)
-            return result[idx]
-
-        active_observed = np.zeros(1)
-        active_observed[0] = target_observed[idx]
-
-        sampler = lambda : np.random.choice(n, size=(n,), replace=True)
-
-        target_sampler = mv.setup_target(active_target, active_observed)
-        target_scaling = 5 * np.linalg.svd(target_sampler.target_transform[0][0])[1].max()**2# should have something do with noise scale too
-
-        test_stat = lambda x: x[0]
-        pval = target_sampler.hypothesis_test(test_stat, active_observed, burnin=10000, ndraw=8000, stepsize=1./crude_target_scaling(target_sampler)) # twosided by default
-        pvalues.append(pval)
-
-        # true selected
-
-        def active_target(indices):
-            result = boot_target(indices)
-            return np.hstack([result[idx], result[nactive:]])
-
-        active_observed = np.zeros_like(active_target(range(n)))
-        active_observed[0] = target_observed[idx]
-        active_observed[1:] = target_observed[nactive:]
-
-        target_sampler = mv.setup_target(active_target, active_observed, target_set=[0])
-        target_scaling = 5 * np.linalg.svd(target_sampler.target_transform[0][0])[1].max()**2 # should have something do with noise scale too
-
-        test_stat = lambda x: x[0]
-        pval = target_sampler.hypothesis_test(test_stat, active_observed, burnin=10000, ndraw=8000, stepsize=1./crude_target_scaling(target_sampler)) # twosided by default
-        pvalues.append(pval)
-
-        return pvalues
